huge refactor!
